Move OBJ object upload e2e test to multicluster directory
